requirements.txt
sqlalchemy
python-dotenv
psycopg2-binary
sqlalchemy-migrate
protobuf
tqdm
